+++ lib/src/delegate.dart
@@ -1,7 +1,7 @@
 import 'package:app_state/app_state.dart';
 import 'package:enum_map/enum_map.dart';
 import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/widgets.dart';
+import 'package:flutter/material.dart';
 
 import 'notifiers/settings.dart';
 import 'router/page_factory.dart';
@@ -19,7 +19,7 @@ abstract class QuickDelegate<T extends Enum> {
   List<Widget> addSpacing(List<Widget> widgets) =>
       widgets.intersperse(spacing).toList(growable: false);
 
-  final authProviders = [
+  final authProviderIds = [
     GoogleAuthProvider.PROVIDER_ID,
     AppleAuthProvider.PROVIDER_ID,
   ];
@@ -44,6 +44,14 @@ abstract class QuickDelegate<T extends Enum> {
   Widget getMenu() => QuickTopMenuWidget(
         delegate: this,
       );
+
+  Icon getTabIcon(T tab) {
+    return switch (tab.name) {
+      'home' => const Icon(Icons.home),
+      'settings' => const Icon(Icons.settings),
+      _ => const Icon(Icons.circle_outlined),
+    };
+  }
 
   void goHome() {
     pageStacks.setCurrentStackKey(tabs.first.name);